requirements.txt
Django==5.2.3
djangorestframework==3.16.0
drf_spectacular==0.28.0
django-crispy-forms==2.4
crispy-bootstrap5==2025.6
